fix: `tr_torrent::seed_ratio_` should be double (#8084)

+++ libtransmission/torrent.h
@@ -1409,7 +1409,7 @@ private:
     time_t seconds_seeding_before_current_start_ = 0;
 
     float verify_progress_ = -1.0F;
-    float seed_ratio_ = 0.0F;
+    double seed_ratio_ = 0.0;
 
     tr_announce_key_t announce_key_ = tr_rand_obj<tr_announce_key_t>();
